VCST-5361: organize benchmark classes into per-suite folders and align namespaces
