+++ benchmarks/VirtoCommerce.Platform.Benchmark/ArrayVsList/Something.cs
@@ -1,0 +1,8 @@
+using System;
+
+namespace VirtoCommerce.Platform.Benchmark.ArrayVsList;
+
+public class Something
+{
+    public string Id { get; set; } = Guid.NewGuid().ToString();
+}
