--- benchmarks/VirtoCommerce.Platform.Benchmark/Something.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace ArrayVsListBench
-{
-    public class Something
-    {
-        public string Id { get; set; } = Guid.NewGuid().ToString();
-    }
-}
